Move encryption services to modules
